rework(pagination): styled pagination to squares with min-width; changed shown quantity of pages to 3; remove pagination_button component; moved pagination from folder;
