--- src/components/composite/zPagination/partial/PaginationButton.vue
@@ -1,15 +1,0 @@
-<template >
-  <button class="flex items-center px-2.5 py-2 uppercase" >
-    <slot />
-  </button>
-</template>
-
-<script>
-export default {
-  name: 'PaginationButton',
-}
-</script>
-
-<style lang="scss" scoped>
-
-</style>
